intial consolidation
